finish world.create

+++ Frent/Core/Component.cs
@@ -42,7 +42,7 @@ public static class Component<T>
         }
     }
 
-    internal static /*readonly*/ nint SparseSetComponentIndex => throw new NotImplementedException();
+    internal static /*readonly*/ int SparseSetComponentIndex => throw new NotImplementedException();
 
     internal static readonly UpdateMethodData[] UpdateMethods;
     internal static readonly ComponentBufferManager<T> BufferManagerInstance;
